+++ src/main/groovy/life/qbic/datamodel/dtos/business/services/ProteomicAnalysis.groovy
@@ -1,6 +1,7 @@
 package life.qbic.datamodel.dtos.business.services
 
 import groovy.transform.EqualsAndHashCode
+import life.qbic.datamodel.dtos.business.ProductCategory
 import life.qbic.datamodel.dtos.business.ProductId
 import life.qbic.datamodel.dtos.business.facilities.Facility
 
@@ -21,12 +22,12 @@ class ProteomicAnalysis extends AtomicProduct {
     * @param description The description of what the product is about.
     * @param unitPrice The price in € per unit
     * @param unit The product unit
-    * @param runningNumber Number used in conjunction with ProductType{@link life.qbic.datamodel.dtos.business.services.ProductType} to identify product
+    * @param runningNumber Number used in conjunction with {@link ProductCategory} to identify product
     * @deprecated since 2.5.0
     */
     @Deprecated
     ProteomicAnalysis(String name, String description, double unitPrice, ProductUnit unit, String runningNumber) {
-        super(name, description, unitPrice, unit, new ProductId(ProductType.PROTEOMIC.toString(), runningNumber))
+        super(name, description, unitPrice, unit, new ProductId(ProductCategory.PROTEOMIC.getAbbreviation(), runningNumber))
     }
 
     /**
@@ -38,13 +39,13 @@ class ProteomicAnalysis extends AtomicProduct {
      * @param description The description of what the product is about.
      * @param unitPrice The price in € per unit
      * @param unit The product unit
-     * @param runningNumber Number used in conjunction with ProductType{@link life.qbic.datamodel.dtos.business.services.ProductType} to identify product
+     * @param runningNumber Number used in conjunction with {@link ProductCategory} to identify product
      *
      * @deprecated 2.11.0
      */
     @Deprecated
     ProteomicAnalysis(String name, String description, double unitPrice, ProductUnit unit, long runningNumber) {
-        super(name, description, unitPrice, unit, new ProductId.Builder(ProductType.PROTEOMIC.toString(), runningNumber).build())
+        super(name, description, unitPrice, unit, new ProductId.Builder(ProductCategory.PROTEOMIC.getAbbreviation(), runningNumber).build())
     }
 
   /**
@@ -57,12 +58,12 @@ class ProteomicAnalysis extends AtomicProduct {
    * @param internalUnitPrice The price in € per unit for internal customers
    * @param externalUnitPrice The price in € per unit for external customers
    * @param unit The product unit
-   * @param runningNumber Number used in conjunction with ProductType{@link life.qbic.datamodel.dtos.business.services.ProductType} to identify product
+   * @param runningNumber Number used in conjunction with {@link ProductCategory} to identify product
    * @param serviceProvider The facility providing the service product
    *
    * @since 2.11.0
    */
   ProteomicAnalysis(String name, String description, double internalUnitPrice, double externalUnitPrice, ProductUnit unit, long runningNumber, Facility serviceProvider) {
-    super(name, description, internalUnitPrice, externalUnitPrice, unit, new ProductId.Builder(ProductType.PROTEOMIC.toString(), runningNumber).build(), serviceProvider)
+    super(name, description, internalUnitPrice, externalUnitPrice, unit, new ProductId.Builder(ProductCategory.PROTEOMIC.getAbbreviation(), runningNumber).build(), serviceProvider)
   }
 }